refactor(json-patch-diff): 💡 rename Diff to JsonPatchDiff
